+++ Main.cpp
@@ -1351,7 +1351,7 @@ bool PlayIntroTitle() {
         {
             if (sectionSwitch)
             {
-				duration = 1.0F * g_SettingsMan.GetMenuTransitionSpeed();
+				duration = 1.0F * g_SettingsMan.GetMenuTransitionDurationMultiplier();
                 sectionSwitch = false;
                 scrollOffset.m_Y = preMenuYOffset;
 
@@ -1420,7 +1420,7 @@ bool PlayIntroTitle() {
         {
             if (sectionSwitch)
             {
-                duration = 2.0F * g_SettingsMan.GetMenuTransitionSpeed();
+                duration = 2.0F * g_SettingsMan.GetMenuTransitionDurationMultiplier();
                 sectionSwitch = false;
 
                 // Reset the Scenario menu
@@ -1446,7 +1446,7 @@ bool PlayIntroTitle() {
                 // Black fade
                 clear_to_color(pFadeScreen, 0);
 
-				duration = 1.0F * g_SettingsMan.GetMenuTransitionSpeed();
+				duration = 1.0F * g_SettingsMan.GetMenuTransitionDurationMultiplier();
                 sectionSwitch = false;
             }
 
@@ -1504,7 +1504,7 @@ bool PlayIntroTitle() {
         {
             if (sectionSwitch)
             {
-				duration = 2.0F * g_SettingsMan.GetMenuTransitionSpeed();
+				duration = 2.0F * g_SettingsMan.GetMenuTransitionDurationMultiplier();
                 sectionSwitch = false;
 
                 // Play the campaign music with Meta sound start
@@ -1527,7 +1527,7 @@ bool PlayIntroTitle() {
                 // Black fade
                 clear_to_color(pFadeScreen, 0);
 
-				duration = 1.0F * g_SettingsMan.GetMenuTransitionSpeed();
+				duration = 1.0F * g_SettingsMan.GetMenuTransitionDurationMultiplier();
                 sectionSwitch = false;
             }
 
@@ -1576,7 +1576,7 @@ bool PlayIntroTitle() {
         {
             if (sectionSwitch)
             {
-				duration = 2.0F * g_SettingsMan.GetMenuTransitionSpeed();
+				duration = 2.0F * g_SettingsMan.GetMenuTransitionDurationMultiplier();
                 sectionSwitch = false;
             }
 
@@ -1592,7 +1592,7 @@ bool PlayIntroTitle() {
             {
                 // Black fade
                 clear_to_color(pFadeScreen, 0x00000000);
-				duration = 1.5F * g_SettingsMan.GetMenuTransitionSpeed();
+				duration = 1.5F * g_SettingsMan.GetMenuTransitionDurationMultiplier();
                 sectionSwitch = false;
             }
 
@@ -1618,7 +1618,7 @@ bool PlayIntroTitle() {
             {
                 // White fade
                 clear_to_color(pFadeScreen, 0x00000000);
-				duration = 1.5F * g_SettingsMan.GetMenuTransitionSpeed();
+				duration = 1.5F * g_SettingsMan.GetMenuTransitionDurationMultiplier();
                 sectionSwitch = false;
             }